requirements.txt
Flask==3.1.0
marshmallow==4.0.0
gunicorn==23.0.0
python-dotenv==1.1.0